feat: Add webpack-dev-server to serve files
This now replaces http-server in the project.

Note that the webpack-dev-server also takes care of running the webpack
process and creating the bundle file, but unlike just running 'webpack',
the webpack-dev-server serves the files from memory instead of from the
file system.

Webpack-dev-server also has automatic reloading built in, so when you
make a change to a source file (that webpack manages), you will not have
to refresh the browser manually.