feat: Add a custom webpack loader to print to the console
It is possible to debug a webpack build in chrome by executing the
following:

node --inspect-brk node_modules/webpack/bin/webpack.js --env production

The '--inspect-brk' forces the node application to break on the first
line of execution.  Node doesn't know how to map into the node_modules
folder so that is done manually to the webpack.js file.  And the
'--env production' is required by my webpack configuration to specify
the build type.

The rest of the process of debugging a node application in chrome can be
made substantially easier with the Node --inspect Manager chrome
extension.  This will automatically launch the chrome dev tools attached
to the running node application.

Debugging in chrome is very useful for a more feature filled console.

---

The loader is called a 'tee-loader' because it doesn't actually perform
any transformation on the code like any other loader would do.  It
passes the source straight across, and just dumps it out to the console
first.