feat: Only run babel-loader for production
This has the potential to really speed up the repeated development
builds.

In order to do this I had to create a "no operation loader" to take the
babel-loader's place because simply undefined wasn't working.  A better
way to handle this might be webpack-merge, but I like how this maintains
a single config for development and production, and let's you see the
differences easily.